+++ waggle-dance-extensions/src/main/java/com/hotels/bdp/waggledance/extensions/client/ratelimit/RateLimitingInvocationHandler.java
@@ -24,18 +24,18 @@ import org.apache.hadoop.hive.metastore.HiveMetaStore.HMSHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.google.common.collect.Sets;
-import com.hotels.bdp.waggledance.client.CloseableThriftHiveMetastoreIface;
-import com.hotels.bdp.waggledance.server.WaggleDanceServerException;
-
 import io.github.bucket4j.Bucket;
 import io.github.bucket4j.ConsumptionProbe;
 import io.micrometer.core.instrument.MeterRegistry;
 
+import com.google.common.collect.Sets;
+
+import com.hotels.bdp.waggledance.client.CloseableThriftHiveMetastoreIface;
+import com.hotels.bdp.waggledance.server.WaggleDanceServerException;
+
 class RateLimitingInvocationHandler implements InvocationHandler {
   private static Logger log = LoggerFactory.getLogger(RateLimitingInvocationHandler.class);
 
-  static final String METRIC_BASE_NAME = "com.hotels.bdp.waggledance.extensions.client.ratelimit";
   static final String UNKNOWN_USER = "_UNKNOWN_USER_";
   private static final Set<String> IGNORABLE_METHODS = Sets.newHashSet("isOpen", "close", "set_ugi", "flushCache");
   private String metastoreName;
@@ -76,7 +76,7 @@ class RateLimitingInvocationHandler implements InvocationHandler {
     if (shouldProceedWithCall(method)) {
       return doRealCall(client, method, args);
     } else {
-      meterRegistry.counter(Metrics.EXCEEDED.getMetricName()).increment();
+      meterRegistry.counter(RateLimitMetrics.EXCEEDED.getMetricName()).increment();
       log.info("User '{}' made too many requests.", user);
       // HTTP status would be 429, so using same for Thrift.
       throw new WaggleDanceServerException("[STATUS=429] Too many requests.");
@@ -92,7 +92,7 @@ class RateLimitingInvocationHandler implements InvocationHandler {
               method.getName(), HMSHandler.getThreadLocalIpAddress(), probe.getRemainingTokens(), metastoreName);
       return probe.isConsumed();
     } catch (Exception e) {
-      meterRegistry.counter(Metrics.ERRORS.getMetricName()).increment();
+      meterRegistry.counter(RateLimitMetrics.ERRORS.getMetricName()).increment();
       if (log.isDebugEnabled()) {
         log.error("Error while processing rate limit for: User:{}, method:{}", user, method.getName(), e);
       } else {
@@ -102,7 +102,7 @@ class RateLimitingInvocationHandler implements InvocationHandler {
       }
       return true;
     } finally {
-      meterRegistry.counter(Metrics.CALLS.getMetricName()).increment();
+      meterRegistry.counter(RateLimitMetrics.CALLS.getMetricName()).increment();
     }
   }
